fix energy bug in number-mc: we were double-counting some atoms
The issue is that with the fast new optcell module, when you have a
very small cell a given atom can show up more than once on the list of
possible nearest neighbors, with shifts of a lattice vector.  This
ensures that we always have the closest version, but means that if we
then try to shift the position to the closest image, we can see the
same atom several times, which of course is wrong (and is what we were
doing).