new scheme for saving movies with "binning" algorithm
This new method should have some advantages.  Instead of storing the
entire movie data in the checkpoint, instead we just store a whole
bunch of checkpoint files.  This means that reading the movie data is
identical to reading the "latest" data.  So once we know how to read
the "latest" data, we can read the movie data identically.